Fix integration test imports. Fix accidental rename issues. Lint.
