+++ openff/evaluator/tests/test_attributes/test_attributes.py
@@ -1,5 +1,5 @@
 """
-Units tests for openff-evaluator.attributes
+Units tests for openff.evaluator.attributes
 """
 import json
