+++ openff/evaluator/tests/test_datasets/test_thermoml.py
@@ -1,5 +1,5 @@
 """
-Units tests for openff-evaluator.datasets
+Units tests for openff.evaluator.datasets
 """
 import pint
 import pytest
